pyproject.toml
[build-system]
requires = ["hatchling>=1.27", "hatch-vcs>=0.4"]
build-backend = "hatchling.build"

[project]
name = "humemdb"
dynamic = ["version"]
description = "Embedded multi-model database orchestration over SQLite, DuckDB, Cypher, and exact and ANN vector search"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
  { name = "HumemAI" },
]
dependencies = [
  "antlr4-python3-runtime>=4.13,<5",
  "duckdb>=1.4,<2",
  "lancedb>=0.27,<1",
  "numpy>=2,<3",
  "sqlglot[c]>=30,<31",
  "threadpoolctl>=3.6,<4",
]
requires-python = ">=3.10"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Database",
]
keywords = ["database", "humemdb", "memory"]

[project.urls]
Homepage = "https://github.com/humemai/humemdb"
Documentation = "https://docs.humem.ai/humemdb/"
Repository = "https://github.com/humemai/humemdb"
Issues = "https://github.com/humemai/humemdb/issues"

[dependency-groups]
docs = [
  "mike>=2.1,<3",
  "mkdocs-macros-plugin>=1.3,<2",
  "mkdocs-material>=9.6,<10",
]

[tool.hatch.version]
source = "vcs"
tag-pattern = "v(?P<version>.*)"

[tool.hatch.version.raw-options]
fallback_version = "0.1.0"
local_scheme = "no-local-version"

[tool.hatch.build.targets.wheel]
packages = ["src/humemdb"]
